test glance protection rules (#4)
* add basic test for glance protected images

related to https://github.com/ChameleonCloud/chi-in-a-box/pull/341

to verify that glance property protections are correctly enabled, we
should have a test for it. To enable it, specify a list of tags in the following config option.
For now, only the first item in the list will be verified, simply asserting that unprivileged users cannot "set" the tag.

```
[image]
image_protected_properties = "protected-tag1,protected_tag_2,"
```