add initial container tests (#6)
This adds an initial end-to-end test for container launches on chi@edge

zun-tempest-plugin is added as a dependency in order to reuse the zun API client within it. However, as the zun plugin does not have any scenario tests, we implement our own scenario test base class.

The major behavior differences between chi@edge and "upstream" zun are:

a reservation hint must be passed in order to launch containers
containers move from created to running automatically, instead of needing to be started manually
isolated networks are not supported
block storage is not supported
snapshot is not supported
"capsules" are not supported