start using toggles to skip blazar plugins (#10)
fix a typo and start using the reservation-features-enabled group
to toggle which tests to run.

initially, this toggles tests of the host_plugin on baremetal sites.
TODO: configure toggles for edge device tests and reservable instance
tests.

the service_available.ironic config toggle is only present if the ironic-tempest-plugin is installed
additionally, there may be cases where blazar host reservation isn't enabled, but ironic is.